+++ tests/conftest.py
@@ -1,5 +1,4 @@
 import os
-import shutil
 
 import pytest
 
@@ -8,19 +7,27 @@ from featherstore import _metadata
 
 from .fixtures import DB_PATH, MD_NAME, STORE_NAME
 from .fixtures.database import setup_incompatible_database
+from .fixtures.misc import paths as path_ops
+
+
+@pytest.fixture
+def paths():
+    return path_ops
 
 
 @pytest.fixture(scope="function", name="store")
-def setup_db():
-    with SetupDB() as store:
+def setup_db(paths):
+    with SetupDB(paths) as store:
         yield store
 
 
 class SetupDB:
+    def __init__(self, paths):
+        self._paths = paths
+
     def __enter__(self):
         # Setup
-        if os.path.exists(DB_PATH):
-            shutil.rmtree(DB_PATH, ignore_errors=False)
+        self._paths.rmtree(DB_PATH)
         fs.create_database(DB_PATH, connect=False)
         fs.connect(DB_PATH)
         fs.create_store(STORE_NAME)
@@ -34,20 +41,18 @@ class SetupDB:
                 store.drop_table(table)
             fs.drop_store(store_name, warnings="ignore")
         fs.disconnect()
-        shutil.rmtree(DB_PATH, ignore_errors=False)
+        self._paths.rmtree(DB_PATH)
 
 
 @pytest.fixture(scope="function")
-def create_db():
+def create_db(paths):
     # Setup
-    if os.path.exists(DB_PATH):
-        shutil.rmtree(DB_PATH)
+    paths.rmtree(DB_PATH)
     fs.create_database(DB_PATH, connect=False)
     # Test
     yield
     # Teardown
-    if os.path.exists(DB_PATH):
-        shutil.rmtree(DB_PATH)
+    paths.rmtree(DB_PATH)
 
 
 @pytest.fixture(scope="function")
@@ -62,38 +67,38 @@ def connect_to_db():
 
 
 @pytest.fixture(scope="function")
-def empty_directory():
-    if os.path.exists(DB_PATH):
-        shutil.rmtree(DB_PATH)
+def empty_directory(paths):
+    paths.rmtree(DB_PATH)
     os.makedirs(DB_PATH)
     yield
-    shutil.rmtree(DB_PATH)
+    paths.rmtree(DB_PATH)
 
 
 @pytest.fixture(scope="function")
-def incompatible_database(request):
-    if os.path.exists(DB_PATH):
-        shutil.rmtree(DB_PATH)
+def incompatible_database(request, paths):
+    paths.rmtree(DB_PATH)
     setup_incompatible_database(DB_PATH, request.param)
     yield request.param
-    shutil.rmtree(DB_PATH)
+    paths.rmtree(DB_PATH)
 
 
 @pytest.fixture(scope="function", name="metadata")
-def setup_md():
-    with SetupMetadata() as md:
+def setup_md(paths):
+    with SetupMetadata(paths) as md:
         yield md
 
 
 class SetupMetadata:
+    def __init__(self, paths):
+        self._paths = paths
+
     def __enter__(self):
         # Setup
-        if os.path.exists(DB_PATH):
-            shutil.rmtree(DB_PATH, ignore_errors=False)
+        self._paths.rmtree(DB_PATH)
         md = _metadata.Metadata(DB_PATH, MD_NAME)
         md.create()
         return md
 
     def __exit__(self, exception_type, exception_value, traceback):
         # Teardown
-        shutil.rmtree(DB_PATH, ignore_errors=False)
+        self._paths.rmtree(DB_PATH)